cleanup premerge
