+++ mani_skill/envs/tasks/tabletop/pick_soda_from_cabinet.py
@@ -71,7 +71,7 @@ class PickSodaFromCabinetEnv(BaseEnv):
         # self.cabinet_scene = RoboCasaSceneBuilder(
         #     env=self, init_robot_base_pos=sapien.Pose(p=[4, -0.6, 0.94], q=[ 0.7071, 0, 0, 0.7071]) )
         # self.cabinet_scene.build(build_config_idxs=[1])
-        
+
         # If you previously built the full robocasa scene, skip it and use this:
         # programmatic open cabinet only:
         # size is width, depth, height (meters)
@@ -94,12 +94,12 @@ class PickSodaFromCabinetEnv(BaseEnv):
         # If environment uses multiple envs, repeat build for each environment index you care about.
         # Optionally keep a handle:
         self.open_cabinet = built
-        self.soda = self.load_glb_as_actor(self.scene, 
-                                             "/home/prajwal-vijay/Downloads/ManiSkill-main/mani_skill/assets/place_soda_in_cabinet/opened_soda_can.glb",
-                                            sapien.Pose(p=[0.055, -0.158, 0.1], q=[0.854,0.471,0.212,0.068]),
-                                            name="soda_can",
-                                            scale=[0.04,0.04,0.04],
-                                            type="dynamic")
+        self.soda = self.load_glb_as_actor(self.scene,
+            os.path.join(os.path.dirname(__file__), '../../../assets/place_soda_in_cabinet/opened_soda_can.glb'),
+            sapien.Pose(p=[0.055, -0.158, 0.1], q=[0.854,0.471,0.212,0.068]),
+            name="soda_can",
+            scale=[0.04,0.04,0.04],
+            type="dynamic")
 
     @staticmethod
     def load_glb_as_actor(scene, glb_file_path, pose, name, scale, type="static"):
@@ -112,7 +112,6 @@ class PickSodaFromCabinetEnv(BaseEnv):
             actor = builder.build_dynamic(name)
         else:
             actor = builder.build_static(name)
-        print(f"{name} imported successfully")
         return actor
 
     def _initialize_episode(self, env_idx: torch.Tensor, options: dict):
