+++ mani_skill/examples/motionplanning/panda/run.py
@@ -11,7 +11,7 @@ from tqdm import tqdm
 import os.path as osp
 from mani_skill.utils.wrappers.record import RecordEpisode
 from mani_skill.trajectory.merge_trajectory import merge_trajectories
-from mani_skill.examples.motionplanning.panda.solutions import solvePushCube, solvePickCube, solveStackCube, solvePegInsertionSide, solvePlugCharger, solvePullCubeTool, solveLiftPegUpright, solvePullCube, solveDrawTriangle, solveDrawSVG, solvePlaceSphere,solveOpenDrawer,solveRaiseCube
+from mani_skill.examples.motionplanning.panda.solutions import solvePushCube, solvePickCube, solveStackCube, solvePegInsertionSide, solvePlugCharger, solvePullCubeTool, solveLiftPegUpright, solvePullCube, solveDrawTriangle, solveDrawSVG, solvePlaceSphere,solveOpenDrawer,solveRaiseCube, solvePlaceBookInShelf, solveHangClothingFrameOnPole, solvePickSodaFromCabinet, solveRotateArrow, solveScoopParticles
 from mani_skill.envs.distraction_set import DISTRACTION_SETS
 
 MP_SOLUTIONS = {
@@ -33,6 +33,12 @@ MP_SOLUTIONS = {
     "OpenDrawer-v1": solveOpenDrawer,               # new
     "PushCube-v2": solvePushCube,                   # new
     "StackCube-v2": solveStackCube,                 # new
+
+    "PlaceBookInShelf-v1": solvePlaceBookInShelf,
+    "HangClothingFrameOnPole-v1": solveHangClothingFrameOnPole,
+    "PickSodaFromCabinet-v1": solvePickSodaFromCabinet,
+    "RotateArrow-v1": solveRotateArrow,
+    "ScoopParticles-v1": solveScoopParticles,
 }
 
 """
@@ -46,6 +52,7 @@ python mani_skill/examples/motionplanning/panda/run.py \
     --num-procs 1 \
     --reward-mode "sparse" \
     --random-seed \
+    --save-video \
     --vis
 """
