+++ mani_skill/examples/motionplanning/panda/solutions/place_book_in_shelf.py
